+++ src/main.hpp
@@ -13,10 +13,10 @@
 
 
 void compute(
-        int16_t* origin, uint8_t len,
-        std::string text, void (*sort)(int16_t* , uint8_t)
+        int16_t* origin, uint32_t len,
+        std::string text, void (*sort)(int16_t* , uint32_t)
 );
-void action(int16_t* origin, uint8_t len);
+void action(int16_t* origin, uint32_t len);
 
 int32_t main(int32_t argc, char** argv);
