+++ src/srt/bubble.hpp
@@ -3,6 +3,6 @@
 
 #include <cstdint>
 
-void bubbleSort(int16_t* point, uint8_t len);
+void bubbleSort(int16_t* point, uint32_t len);
 
 #endif
