+++ src/srt/heap.hpp
@@ -5,7 +5,7 @@
 #include <cmath>
 #include <cstdint>
 
-void heapify(int16_t* point, uint8_t len, uint8_t pos);
-void heapSort(int16_t* point, uint8_t len);
+void heapify(int16_t* point, uint32_t len, uint32_t pos);
+void heapSort(int16_t* point, uint32_t len);
 
 #endif
