+++ src/srt/merge.cpp
@@ -1,22 +1,22 @@
 #include "merge.hpp"
 
-void merge(int16_t* point, uint8_t lpos, uint8_t middle, uint8_t rpos) {
-    uint8_t llen = middle - lpos + 1;
-    uint8_t rlen = rpos - middle;
+void merge(int16_t* point, uint32_t lpos, uint32_t middle, uint32_t rpos) {
+    uint32_t llen = middle - lpos + 1;
+    uint32_t rlen = rpos - middle;
 
     int16_t* llist = new int16_t[llen];
     int16_t* rlist = new int16_t[rlen];
 
-    for (uint8_t idx = 0; idx < llen; idx++) {
+    for (uint32_t idx = 0; idx < llen; idx++) {
         llist[idx] = point[lpos + idx];
     }
-    for (uint8_t idx = 0; idx < rlen; idx++) {
+    for (uint32_t idx = 0; idx < rlen; idx++) {
         rlist[idx] = point[middle + 1 + idx];
     }
 
-    uint8_t ll = 0;
-    uint8_t rr = 0;
-    uint8_t pp = lpos;
+    uint32_t ll = 0;
+    uint32_t rr = 0;
+    uint32_t pp = lpos;
 
     while (ll < llen && rr < rlen) {
         if (llist[ll] <= rlist[rr]) {
@@ -45,9 +45,9 @@ void merge(int16_t* point, uint8_t lpos, uint8_t middle, uint8_t rpos) {
     delete [] rlist;
 }
 
-void mergeSort(int16_t* point, uint8_t lpos, uint8_t rpos) {
+void mergeSort(int16_t* point, uint32_t lpos, uint32_t rpos) {
     if (lpos < rpos) {
-        uint8_t middle = floor((lpos + rpos) / 2);
+        uint32_t middle = floor((lpos + rpos) / 2);
         mergeSort(point, lpos, middle);
         mergeSort(point, middle + 1, rpos);
 
@@ -55,6 +55,6 @@ void mergeSort(int16_t* point, uint8_t lpos, uint8_t rpos) {
     }
 }
 
-void mergeSort(int16_t* point, uint8_t len) {
+void mergeSort(int16_t* point, uint32_t len) {
     mergeSort(point, 0, len - 1);
 }
